laird: libnm-core: support for setting multiple phase2-auth and phase2-autheaps
The wpa_supplicant can handle multiple 'auth' and 'autheap' variables
inside of the phase2 property.  These two settings become lists instead
of strings and the appropriate getter/setters are being added. The
functions to verify the settings and check for secrets are updated as well
to traverse the list and validate multiple settings.

Bug 10820