lrd-network-manager: Fix build source selection when building with introspection and without docs
A select group of documentation source files are added to the build sources when
the '--enable-introspection' flag is enabled which should be reserved for only
when the '--enable-gtk-doc' flag is enabled. Therfore, builds with introspection
and without the docs will fail. This change introduces extra if-guards around
the pertinent files in the Makefile.

Bug 22029