lrd-network-manager: Remove dependency on pygobject to build with GObject introspection support
This change fixes the build configuration logic to allow for building
NetworkManager with GObject introspection support enabled without an explicit
dependency on pygobject. pygobject is necessary for generating the documentation
(which is assumed to be the only reason for enabling introspection); however, it
is not necessary if not generating the documenation.

Bug 22029