lrd-network-manager: Add reporting of DHCP Vendor Specific Options (43)
This change adds the contents of DHCP option 43 (Vendor Specific Options or
Vendor Encapsulated Options) to the list of reported DHCP options if the option
is present in the lease returned from a DHCP server.

Option 43 is formatted as a set of hex bytes and must therefore be converted to
printable characters prior to being added to the list of DHCP options. This was
modeled after the dhcp_lease_save() function in sd-dhcp-lease.c which prints
the DHCP options to the lease file.

Bug 22356